fixed #19, fixed #12: moved mockdata files and files under components/tests
