+++ src/components/CharacterDetails/CharacterDetails.tsx
