+++ src/components/CharacterList/CharacterList.test.tsx
@@ -1,6 +1,6 @@
 import { screen, render } from '@testing-library/react'
-import CharacterList from '../CharacterList'
-import { modifiedCharacters } from './mocks/modifiedCharacters'
+import CharacterList from './CharacterList'
+import { modifiedCharacters } from '../../api/mocks/modifiedCharacters'
 
 describe('CharacterList', () => {
   test('renders a list of characters', () => {
