+++ src/components/SpeciesItem/SpeciesItem.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react'
 import { Tabs, TabPanels } from '@chakra-ui/react'
-import SpeciesItem from '../SpeciesItem'
-import { modifiedCharacters } from './mocks/modifiedCharacters'
+import SpeciesItem from './SpeciesItem'
+import { modifiedCharacters } from '../../api/mocks/modifiedCharacters'
 
 describe('SpeciesItem', () => {
   test('renders a species list', () => {
