+++ src/mocks/handlers.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw'
 import { RowCharacter } from '../api/starWars-types'
-import { characterMocks } from './characterMocs'
+import { characterMocks } from '../api/mocks/characterMocs'
 
 export const handlers = [
   rest.get('https://akabab.github.io/starwars-api/api/all.json', (_, res, ctx) =>
